refactor: extract trait for queueable events represented by JSON objects
